backend/requirements.txt
alembic==1.18.4
blinker==1.9.0
boto3==1.42.68
botocore==1.42.68
click==8.3.1
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2026.2.0
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
setuptools==82.0.1
six==1.17.0
SQLAlchemy==2.0.48
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
